+++ controller/auth.js
@@ -171,49 +171,6 @@ const logout = catchAsync(async (req, res, next) => {
 })
 
 /*
-  修改密碼 PATCH /reset-password
-*/
-const resetPassword = catchAsync(async (req, res, next) => {
-  const user = req.user
-  const { password, confirmPassword } = req.body
-  if (!password || !confirmPassword) {
-    return next(
-      new AppError({
-        statusCode: ApiState.FIELD_MISSING.statusCode,
-        status: ApiState.FIELD_MISSING.status,
-        message: '新密碼及再次確認新密碼為必填項目',
-      }),
-    )
-  }
-  if (!checkPassword(req.body.password)) {
-    return next(
-      new AppError({
-        statusCode: ApiState.FIELD_MISSING.statusCode,
-        status: ApiState.FIELD_MISSING.status,
-        message: '密碼格式錯誤，需包含至少一個英文字與數字，密碼八碼以上',
-      }),
-    )
-  }
-  if (password !== confirmPassword) {
-    return next(
-      new AppError({
-        statusCode: ApiState.FAIL.statusCode,
-        status: ApiState.FAIL.status,
-        message: '新密碼及再次確認新密碼不一致',
-      }),
-    )
-  }
-  const newPassword = hashPassword(password)
-  const editUser = await User.findByIdAndUpdate(user?.id, { password: newPassword }, { returnDocument: 'after', runValidators: true })
-  if (!editUser) {
-    return next(
-      new AppError(ApiState.UPDATE_FAILED),
-    )
-  }
-  successHandle({ res, message: '修改密碼成功', data: editUser })
-})
-
-/*
   驗證token GET /check
 */
 const checkToken = catchAsync(async (req, res, next) => {
@@ -295,7 +252,7 @@ module.exports = {
   login,
   signup,
   logout,
-  resetPassword,
   checkToken,
   isAuth,
+  checkPassword,
 }
